requirements.txt
streamlit
pandas
openpyxl
plotly
crewai
langchain_google_genai
python-dotenv
tabulate
